+++ src/main/java/algogather/api/controller/notification/NotificationController.java
@@ -3,6 +3,7 @@ package algogather.api.controller.notification;
 import algogather.api.domain.user.UserAdapter;
 import algogather.api.dto.api.ApiResponse;
 import algogather.api.dto.notification.NotificationListResponseDto;
+import algogather.api.dto.notification.NotificationReadRequestDto;
 import algogather.api.dto.programminglanguage.ProgrammingLanguageListResponseDto;
 import algogather.api.service.notification.NotificationService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -15,9 +16,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
 
 @RestController
 @RequiredArgsConstructor
@@ -53,4 +54,17 @@ public class NotificationController {
 
         return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.sucessWithDataAndMessage(existsNew, "새 알림 존재 여부를 정상적으로 조회했습니다."));
     }
+
+//    @Operation(summary = "알림 읽기 API", description = "알림 읽기 API입니다.")
+//    @ApiResponses(value = {
+//            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "성공"),
+//            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "400", description = "실패", content = @Content(schema = @Schema(implementation = ApiResponse.class))),
+//            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "401", description = "인증 실패", content = @Content(schema = @Schema(implementation = ApiResponse.class)))
+//    })
+//    @PostMapping("/read")
+//    public ResponseEntity<ApiResponse<Boolean>> readNotification(@Valid @RequestBody NotificationReadRequestDto notificationReadRequestDto, @Parameter(hidden = true) @AuthenticationPrincipal UserAdapter userAdapter) {
+//
+//
+//        return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.sucessWithDataAndMessage(, "새 알림 존재 여부를 정상적으로 조회했습니다."));
+//    }
 }
